manual merge of MPIDataMining branch
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2863 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/algorithm/DMSystemMatrixSPVectorizedIdentityMPI.hpp
@@ -100,7 +100,7 @@ namespace sg {
          * @param blocksize blocksize
          *
          */
-        void calcDistribution(int totalSize, int* sizes, int* offsets, size_t blocksize);
+        void calcDistribution(size_t totalSize, int* sizes, int* offsets, size_t blocksize);
 
     };
